WareHouse/idp_SCALAR-CHATDEV_20231012015123/requirements.txt
jinja2
python-gitlab
